requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-core
langchain-google-genai
google-generativeai
